pyproject.toml
[project]
name = "autoagents-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.2.45",
    "pyyaml>=6.0.0",
    "crewai",
    "crewai[tools]",
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "seaborn>=0.13.2",
]
